Move SegModel to src/utils
